Fix for Cardboard storage of ints into float openings array for texture clipping purposes (yeah this should have been in trunk, so sue me).

+++ source/r_defs.h
@@ -526,7 +526,7 @@ typedef struct drawseg_s
 
   float *sprtopclip, *sprbottomclip;
   // SoM: this still needs to be int
-  int   *maskedtexturecol;
+  float *maskedtexturecol;
 
   fixed_t viewx, viewy, viewz;
 } drawseg_t;
